+++ CHANGELOG.md
@@ -84,6 +84,7 @@
     + UIView methods: `-addTapGestureHandler:`, `-addLongPressGestureHandler:`, `-all...GestureRecognizers`, `-removeAll...GestureRecognizers`, `-setCornerRadius:borderWidth:borderColor:`
     + UIBarButtonItem block methods
     + `-[UIViewController currentVisibleViewController]`
+    + `-[UITabBarController setBadgeValue:forTabBarItemAtIndex:]`
 - Refactored:
     + `NSObject (ESAutoCoding)`
     + `-queryDictionary` of `NSURL`/`NSString` to use the `NSURLComponents` API